finish C++ - Module 01

+++ C++ Module 01/ex01/zombieHorde.cpp
@@ -9,7 +9,7 @@ Zombie *zombieHorde(int N, std::string name)
 
     horde = new Zombie[N];
 
-    for (int i = 0; i < N; i++)
+    for (short i = 0; i < N; i++)
         horde[i].setName(name);
 
     return (horde);
